+++ pccommon/pccommon/middleware.py
@@ -6,14 +6,13 @@ from fastapi.applications import FastAPI
 from starlette.middleware.base import BaseHTTPMiddleware
 from starlette.types import Message
 
-from pccommon.logging import request_to_path
-from pccommon.tracing import HTTP_METHOD, HTTP_PATH, HTTP_URL, trace_request
+from pccommon.logging import get_custom_dimensions
+from pccommon.tracing import trace_request
 
 logger = logging.getLogger(__name__)
 
 
 async def handle_exceptions(
-    service_name: str,
     request: Request,
     call_next: Callable[[Request], Awaitable[Response]],
 ) -> Response:
@@ -24,15 +23,7 @@ async def handle_exceptions(
     except Exception as e:
         logger.exception(
             "Exception when handling request",
-            extra={
-                "custom_dimensions": {
-                    "stackTrace": f"{e}",
-                    HTTP_URL: str(request.url),
-                    HTTP_METHOD: str(request.method),
-                    HTTP_PATH: request_to_path(request),
-                    "service": service_name,
-                }
-            },
+            extra=get_custom_dimensions({"stackTrace": f"{e}"}, request),
         )
         raise