+++ pctiler/pctiler/reader.py
@@ -1,3 +1,5 @@
+import logging
+import time
 from typing import Any, Dict, List, Optional, Tuple, Type
 
 import attr
@@ -11,11 +13,15 @@ from rio_tiler.io.base import BaseReader
 from rio_tiler.models import ImageData
 from rio_tiler.mosaic import mosaic_reader
 from titiler.pgstac import mosaic as pgstac_mosaic
-from titiler.pgstac.settings import CacheSettings
 from titiler.pgstac.reader import PgSTACReader
+from titiler.pgstac.settings import CacheSettings
+
 from pccommon.cdn import BlobCDN
 from pccommon.config import get_render_config
+from pctiler.config import get_settings
 from pctiler.reader_cog import CustomCOGReader  # type:ignore
+
+logger = logging.getLogger(__name__)
 
 cache_config = CacheSettings()
 
@@ -40,6 +46,8 @@ class ItemSTACReader(PgSTACReader):
 @attr.s
 class MosaicSTACReader(pgstac_mosaic.CustomSTACReader):
     """Custom version of titiler.pgstac.mosaic.CustomSTACReader)."""
+
+    reader: Type[BaseReader] = attr.ib(default=CustomCOGReader)
 
     def _get_asset_url(self, asset: str) -> str:
         """Validate asset names and return asset's url.
@@ -75,6 +83,8 @@ class PGSTACBackend(pgstac_mosaic.PGSTACBackend):
     def assets_for_tile(
         self, x: int, y: int, z: int, collection: Optional[str] = None, **kwargs: Any
     ) -> List[Dict]:
+        settings = get_settings()
+
         # Require a collection
         if not collection:
             raise HTTPException(
@@ -82,13 +92,39 @@ class PGSTACBackend(pgstac_mosaic.PGSTACBackend):
                 detail="Tile request must contain a collection parameter.",
             )
 
-        # Check that the zoom isn't lower than minZoom
+        ts = time.perf_counter()
         render_config = get_render_config(collection)
-        if render_config and render_config.minzoom and render_config.minzoom > z:
+
+        # Don't render if this collection is unconfigured
+        if not render_config:
             return []
 
+        # Check that the zoom isn't lower than minZoom
+        if render_config.minzoom and render_config.minzoom > z:
+            return []
+
+        # Override items_limit via render config for collection
+        max_items = (
+            render_config.max_items_per_tile or settings.default_max_items_per_tile
+        )
+        asset_kwargs = {**kwargs, "items_limit": max_items}
+
         bbox = self.tms.bounds(morecantile.Tile(x, y, z))
-        return self.get_assets(Polygon.from_bounds(*bbox), **kwargs)
+        assets = self.get_assets(Polygon.from_bounds(*bbox), **asset_kwargs)
+
+        te = time.perf_counter()
+        logger.info(
+            "Perf: Mosaic get assets for tile.",
+            extra={
+                "custom_dimensions": {
+                    "duration": f"{te - ts:0.4f}",
+                    "collection": collection,
+                    "zxy": f"{z}/{x}/{y}",
+                    "count": len(assets),
+                },
+            },
+        )
+        return assets
 
     # override from PGSTACBackend to pass through collection
     def tile(
@@ -123,6 +159,7 @@ class PGSTACBackend(pgstac_mosaic.PGSTACBackend):
                 f"No assets found for tile {tile_z}-{tile_x}-{tile_y}"
             )
 
+        ts = time.perf_counter()
         if reverse:
             mosaic_assets = list(reversed(mosaic_assets))
 
@@ -134,7 +171,7 @@ class PGSTACBackend(pgstac_mosaic.PGSTACBackend):
             ) as src_dst:
                 return src_dst.tile(x, y, z, **kwargs)
 
-        return mosaic_reader(
+        tile = mosaic_reader(
             mosaic_assets,
             _reader,
             tile_x,
@@ -143,3 +180,19 @@ class PGSTACBackend(pgstac_mosaic.PGSTACBackend):
             allowed_exceptions=(TileOutsideBounds, MissingAssets, InvalidAssetName),
             **kwargs,
         )
+
+        te = time.perf_counter()
+
+        logger.info(
+            "Perf: Mosaic read tile.",
+            extra={
+                "custom_dimensions": {
+                    "duration": f"{te - ts:0.4f}",
+                    "collection": collection,
+                    "zxy": f"{tile_z}/{tile_x}/{tile_y}",
+                    "count": len(mosaic_assets),
+                }
+            },
+        )
+
+        return tile