+++ scripts/bin/setup_azurite.py
@@ -10,13 +10,13 @@ from pathlib import Path
 from azure.data.tables import TableServiceClient
 
 import pccommon
-from pccommon.constants import (
-    DEFAULT_CONTAINER_CONFIG_TABLE_NAME,
-    DEFAULT_COLLECTION_CONFIG_TABLE_NAME,
-    DEFAULT_IP_EXCEPTION_CONFIG_TABLE_NAME
-)
 from pccommon.config.collections import CollectionConfig, CollectionConfigTable
 from pccommon.config.containers import ContainerConfig, ContainerConfigTable
+from pccommon.constants import (
+    DEFAULT_COLLECTION_CONFIG_TABLE_NAME,
+    DEFAULT_CONTAINER_CONFIG_TABLE_NAME,
+    DEFAULT_IP_EXCEPTION_CONFIG_TABLE_NAME,
+)
 from pccommon.tables import IPExceptionListTable
 
 TEST_DATA_DIR = Path(pccommon.__file__).parent.parent / "tests" / "data-files"
@@ -47,7 +47,7 @@ def setup_azurite() -> None:
     for table in [
         DEFAULT_CONTAINER_CONFIG_TABLE_NAME,
         DEFAULT_COLLECTION_CONFIG_TABLE_NAME,
-        DEFAULT_IP_EXCEPTION_CONFIG_TABLE_NAME
+        DEFAULT_IP_EXCEPTION_CONFIG_TABLE_NAME,
     ]:
         if table not in tables:
             print(f"~ ~ Creating table {table}...")
@@ -86,7 +86,9 @@ def setup_azurite() -> None:
     ip_config_table = IPExceptionListTable(
         lambda: (
             None,
-            table_service_client.get_table_client(DEFAULT_IP_EXCEPTION_CONFIG_TABLE_NAME),
+            table_service_client.get_table_client(
+                DEFAULT_IP_EXCEPTION_CONFIG_TABLE_NAME
+            ),
         )
     )
     ip_config_table.add_exception("127.0.0.1")